// apps/web/src/api/generated/model/updateUserDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */
import type { UpdateUserDtoRole } from './updateUserDtoRole.ts';

export interface UpdateUserDto {
  /** @maxLength 100 */
  name?: string;
  role?: UpdateUserDtoRole;
  disabled?: boolean;
}