// apps/web/src/api/generated/model/userDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */
import type { UserDtoRole } from './userDtoRole.ts';

export interface UserDto {
  role: UserDtoRole;
  id: string;
  email: string;
  name: string;
  disabled: boolean;
  createdAt: string;
}